Cleaning up

+++ install.sh
@@ -5,10 +5,7 @@
 # includes #
 ############
 
-if [ ! -f ./install_config ];
-then
-    cp ./install_config.diff ./install_config
-fi
+[ ! -f ./install_config ] && cp ./install_config.diff ./install_config
 
 source ./install_config
 source ./colors.sh
@@ -68,7 +65,6 @@ dot_is_installed composer && dot_install composer
 dot_is_installed mycli && dot_install mycli
 dot_is_installed pgcli && dot_install pgcli
 dot_is_installed go && dot_install_func go install_go_binaries
-dot_is_installed php && dot_install php
 dot_is_installed npm && dot_install javascript
 dot_is_installed compton && dot_install compton
 dot_is_installed bat && dot_install bat
